+++ NEWS.md
@@ -4,6 +4,8 @@
 
 * Added functions for printing results from `eval_capture_conditions()`. Captured conditions can be printed as either errors or messages with `captured_condition_as_error()` and `captured_condition_as_message()`, respectively. (#282)
 
+* Added functions `update_ard_fmt_fn()` and `update_ard_stat_label()` to update an ARD's formatting function and statistic label, respectively. (#253)
+
 # cards 0.2.2
 
 ## New Features & Updates